Use PriceFeed and V2Oracle from TokenizedDerivative
Removed the test case for confirming prices at expiry: this use case no
longer makes sense because the PriceFeed doesn't provide a price exactly
at expiry time.

The contract behaves as follows:
1. Regular remgargins: Use PriceFeed.
2. Expiry: Don't update NAV or balances, wait for V2Oracle.
3. Default: Update NAV and balances, kick off V2Oracle request.
   One of these things have to happen for final settlement:
   a) Participants can confirm the PriceFeed price.
   b) Settle at the V2Oracle price

I understand that "confirmation" may be going away if we collapse
investor and provider, but I left it in for now.