Publish BarChart prices using tradeTimestamp rather than currentTime
The BarChart price feed is delayed by 10 minutes, so the prices in the
ManualPriceFeed should reflect the timestamp of the price, not the
timestamp of the upload.

Note that because of a QPS limit for the BarChart API, we avoid making
calls to BarChart in cases where we know an updated price isn't
available.