Separate README into v1 and v0 sections
Not sure if this is a good plan or not:
The v1 section is incomplete but correct for the system as it is today.
As we fully build out the system, we'll continue to expand that section,
copying from the v0 section as needed.

Alternatively, we could also just delete the leftover instructions about v0, leaving only the pointer to the v0 branch.

Issue #424

Signed-off-by: Prasad Tare <prasad.tare@gmail.com>