Lift wrapper div outside of Step components to fix transitions
In some cases, if the `div` element returned by `components/StepN.js`
changes entirely, then the the outer `CSSTransition` doesn't work and
results in the CSS class overrides `step-N-enter-done` never getting
applied to the div that has the CSS class `step`. This results in all of
the content having opacity 0 and being invisible.

One way to address this issue is to make sure that the direct child of
`CSSTransition` is always the same wrapper div.

Signed-off-by: Prasad Tare <prasad.tare@gmail.com>