Decode admin proposals in voter dapp (#880)
Adds a button next to an admin proposal that shows up a dialog explaining the proposal. The explanation text is what the CLI shows: we could try to improve both over time. E.g., interpret the most common admin proposals and show special handling for "setInflationRate".

I tried a few different options, such as a separate table and a tooltip, but this option seemed like the best for now.
Table:
![image](https://user-images.githubusercontent.com/5361377/73959645-9ef13d80-48d7-11ea-990d-d5b526cc6755.png)

Popup:
![image](https://user-images.githubusercontent.com/5361377/73959681-af091d00-48d7-11ea-8ca0-d31e1841679a.png)

Fixes #880 

Signed-off-by: Prasad Tare <prasad.tare@gmail.com>