Fix bug in voter dapp deployment due to build warning
The one weird thing is that this change also applies to the sponsor
dapp, where it isn't necessary.
Options here are to:
1. Set CI=false on the entire `deploy_dapp.sh` invocation for voter dapp
only
2. Parameterize `deploy_dapp.sh`
3. Do nothing

PR #880 imports AbiUtils from the voter dapp, which results in a build
warning that has to be ignored. AFAIK the only way to ignore webpack
build warnings from React is to turn off CI mode.

Signed-off-by: Prasad Tare <prasad.tare@gmail.com>