[P1-L19] Explicitly pass in amount of ERC20 to pay as fees
The previous implementation transferred all the approved tokens, which
is suboptimal because allowances are not considered payments. Note that
this PR doesn't change functionality because the only callers to
`payOracleFeesErc20` were other contracts.

We prefer to have an explicit function rather than the callers simply
transferring tokens because while this implementation of Store.sol
doesn't do anything in this function, we want the StoreInterface to have
an explicit function we can plug in to later.

Signed-off-by: Prasad Tare <prasad.tare@gmail.com>