[P02-H02] Restrict liquidations by min collateralization ratio as well (#1351)
A liquidator can now specify a min and a max for collateralization ratio
of the position they are liquidating.

The idea is to avoid a delayed liquidation ending up liquidating a
position in haircut territory. E.g., if a liquidation takes a long time
to process and the underlying price moves significantly in that time.