fix(allowances): Change default allowance value
UNI token's `approve()` method constrains max allowance to 2^96-1 because their contract uses `uint96` for internal balance and allowance tracking. Presumably this is due to gas optimizations. This max value is about 1/3 less of the MAX_UINT_VAL value, but is large enough that it should work for almost all other collateral types.

Signed-off-by: Nick Pai <npai.nyc@gmail.com>